refactor: split up timesheet pdf components and replace pdf workarounds that are no longer needed
